Claim retry logic and random placement strategy
There was a bad mistake in the claim retry logic that I had missed which
was actually not re-issueing the claim transaction for the same instance
requests but instead simply sleeping and then grabbing the next instance
from the requests queue. This bug has now been fixed and the deadlock
retry-placement-and-claim now happens up to 5 times on the same instance
request.

In addition, the random placement strategy when used with the db filter
strategy was sometimes able to complete each partition's work before
actually filling the partition's available resource providers. This was
due to the way that the db filter's quick random resource provider ID
winnowed the end result set and was exacerbated the more partitions
(scheduler processes) were used. As each partition winnowed the set of
resource provider IDs to a small amount, as partitions filled up, the
chances of the random resource provider ID selected by random.randint()
was greater than any provider in that partition that happened to have
available space increased. We now track these occurrences of not finding
a resource provider when using the db filter and random placement
strategy and retry the placement query when these occur, up to a certain
threshold.

We also add two new placement strategies: random-pack and random-spread.
When this random provider not found occurs, these new strategies fall
back to either pack or spread strategy to fill the remaining available
slots in the partition.

Adds stats to the results struct for tracking the random no found
retries separately from the deadlock retries.